// frontend/taskhub-web/src/app/api/generated/model/workItemListItemDtoPagedResult.ts
/**
 * TaskHub.WebApi
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { WorkItemListItemDto } from './workItemListItemDto';


export interface WorkItemListItemDtoPagedResult { 
    items?: Array<WorkItemListItemDto> | null;
    pageNumber?: number;
    totalPages?: number;
    totalCount?: number;
    readonly hasPreviousPage?: boolean;
    readonly hasNextPage?: boolean;
}